requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
Brotli==1.1.0
cairocffi==1.7.1
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
coverage==7.9.2
cssselect2==0.8.0
ecdsa==0.19.1
fastapi==0.115.14
fonttools==4.58.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown2==2.5.3
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pytest==8.4.1
pytest-cov==6.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
joblib
scikit-learn
sqlalchemy
